CloudFunction/HTTP/requirements.txt
google-api-python-client
pandas==0.24.1
PyMySQL==0.9.3
SQLAlchemy==1.2.17
requests==2.21.0
google-cloud-storage==1.13.2
gcsfs




